requirements.txt
Django>=4.2
python-dotenv>=1.0
google-generativeai>=0.7.0
psycopg2-binary>=2.9